% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getis.cluster.R
\name{getis.cluster}
\alias{getis.cluster}
\title{Getis and Ord's Gi* Cluster and Significance Map}
\usage{
getis.cluster(x, listw, zero.policy = NULL, shp, significant = TRUE,
  ...)
}
\arguments{
\item{x}{variable to create cluster and significance map}

\item{listw}{a neighbours list with spatial weights. From package spdep:
a listw object. Use poly2nb (class nb)
and nb2listw (class listw, nb) from package spdep. Can be any type of listw
object, for instance, rook contiguity (common edge) or queen contiguity (common
edge or common vertex)}

\item{zero.policy}{by default = NULL, if FALSE stop with error for any empty
neighbour sets, if TRUE permit the weights list to be formed with zero-length
weights vectors. Parameter inherited from the spdep package.}

\item{shp}{the spatial dataset: sf or SpatialPolygonsDataFrame (spdep)}

\item{significant}{by default is TRUE, if FALSE the significant map is not created}

\item{...}{other parameters similar to original}
}
\value{
one or two maps
}
\description{
Create the Getis Gi* Cluster Map and the corresponding Significance Map.
Maps are done calculating the Local Gi* (localG - spdep) for each
spatial unit and testing its significance.
}
\details{
Using the function localG (spdep) create the Getis Gi* Cluster Map and the
corresponding Significance Map.
The significance map is done testing the null hypothesis (Ho) of zero spatial
autocorrelation for each spatial unit, then plotting a choropleth map with this
legend values: (Not Significant, p-value=0.05, p-value= 0.01, p-value=0.001,
p-value=0.0001, and Neighborless). Most significant clustered spatial units are
those with p-values smaller than 0.0001. Not significant
clustered spatial units are those with p-values grather than 0.05. Gi* Cluster Map
is done based on the significance map, but the choropleth legend is different (Not
- Significant, High-High, Low-Low, Low-High, High-Low, and Neighborless).
}
\section{Links}{


\enumerate{
   \item \href{https://pysal.readthedocs.io/en/v1.11.0/users/tutorials/autocorrelation.html}{Spatial Autocorrelation}
}
}

\examples{
library(spdep)
columbus <- st_read(system.file("extdata/columbus.shp", package="spdepbi")[1], quiet=TRUE)
col_nbq <- poly2nb(columbus)
a.lw <- nb2listw(col_nbq, style="W")
getis.cluster(columbus$CRIME, a.lw, zero.policy = FALSE, columbus, significant=TRUE)
}
\seealso{
\itemize{
  \item Bivariate Moran's I: \code{\link{moran.bi}}
  \item Plot Bivariate Moran's I: \code{\link{moranbi.plot}}
  \item Bivariate Moran's I Test: \code{\link{moranbi.test}}
  \item Bivariate Local Moran's I and Test: \code{\link{localmoran.bi}}
  \item Create object "nb": \code{\link[spdep]{poly2nb}}
  \item Create object "listw"/"nb": \code{\link[spdep]{nb2listw}}
}
}
