% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{\%:::\%}
\alias{\%:::\%}
\title{Title}
\usage{
pkg \%:::\% fun
}
\arguments{
\item{pkg}{package name}

\item{fun}{function name}
}
\value{
function or variable value
}
\description{
Title
}
\examples{
library(spdep)
'spdep' \%:::\% '.spdepOptions'
}
