% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranbi.plot.R
\name{moranbi.plot}
\alias{moranbi.plot}
\title{Plot the Bivariate Moran's Index}
\usage{
moranbi.plot(x, y, listw, spChk = NULL, labels = NULL, xlab = NULL,
  ylab = NULL, zero.policy = NULL, quiet = NULL, ...)
}
\arguments{
\item{x}{the first variable}

\item{y}{the second variable}

\item{listw}{a neighbours list with spatial weights. From package spdep:
a listw object. Use poly2nb (class nb)
and nb2listw (class listw, nb) from package spdep. Can be any type of listw
object, for instance, rook contiguity (common edge) or queen contiguity (common
edge or common vertex)}

\item{spChk}{by default = NULL. Similar meaning and values than parameter spChk
of \code{\link[spdep]{localmoran}}}

\item{labels}{by default = NULL, place labels (point ID) of potential influential
observations. If NULL or TRUE will place labels, if FALSE will remove.}

\item{xlab}{by default = NULL, change this string value if a different label for
horizontal axis is needed.}

\item{ylab}{by default = NULL, change this string value if a different label for
vertical axis is needed.}

\item{zero.policy}{by default = NULL, if FALSE stop with error for any empty
neighbours sets, if TRUE permit the weights list to be formed with zero-length
weights vectors. Similar meaning and values than parameter zero.policy
of \code{\link[spdep]{localmoran}}}

\item{quiet}{by default = NULL. If FALSE a table with potentially
influential observations will be displayed. The table will show next columns:
dfb.1_, dfb.x, dffit, cov.r, cook.d, and hat. If TRUE, not R command window output
will be displayed.}

\item{...}{other parameters similar to original}
}
\value{
a graphic showing Bivariate Moran's Index behavior
}
\description{
Create a scatter plot graphic with lagged variables, one variable in each axis,
where can be seen the trend of the spatial autocorrelation.
}
\details{
Bivariate spatial autocorrelation can be seen graphically using spatial lags plots
of scaled variables. Method used to create spatial lag values in morambi.plot is
lag.listw, and is explained in pages 257 and s58 of Edzer J. Pebesma (2008) -- Edzer
J. Pebesma, Roger S. Bivand und Virgilio Gómez-Rubio. 2008. “Areal Data and Spatial
Autocorrelation.” In Applied Spatial Data Analysis with R, Online-Ausg., 237–72.
Springer New York. For graphic analysis, Variables need to have the same relative
scale. Beforehand use the 'scale' function of R base to perform the transformation.
In the resulting graphic, trend line depicts the spatial autocorrelation.
Random pattern represented in horizontal trend line means no spatial
autocorrelation. Cluster spatial autocorrelation is represented by a regression line
with positive slope. Dispersed spatial autocorrelation (spatial outliers) is shown by
a regression line with negative slope. In the graphic it is possible to identify
with different symbol and using labels, atypical observations that fall apart the
displayed pattern.
}
\section{Links}{


\enumerate{
   \item \href{https://en.wikipedia.org/wiki/Indicators_of_spatial_association}{Indicators of Spatial Association}
   \item Spatial Autocorrelation \href{http://blogs.oregonstate.edu/geo599spatialstatistics/2016/06/08/spatial-autocorrelation-morans/}{(Moran’s I) Test}
   \item Moran's I \href{https://stats.idre.ucla.edu/r/faq/how-can-i-calculate-morans-i-in-r/}{Test}
   \item \href{http://sphweb.bumc.bu.edu/otlt/MPH-Modules/BS/BS704_Confidence_Intervals/BS704_Confidence_Intervals_print.html}{Confidence Intervals}
}
}

\examples{
library(spdep)
columbus <- st_read(system.file("extdata/columbus.shp", package="spdepbi")[1], quiet=TRUE)
col_nbq <- poly2nb(columbus)
a.lw <- nb2listw(col_nbq, style="W")
CRIME <- as.vector(scale(columbus$CRIME))
INCOME <- as.vector(scale(columbus$INC))
moranbi.plot(CRIME,INCOME,quiet=FALSE,zero.policy=FALSE,listw=a.lw)
}
\seealso{
\itemize{
  \item Bivariate Moran's I: \code{\link{moran.bi}}
  \item Plot Bivariate Moran's I: \code{\link{moranbi.plot}}
  \item Bivariate Moran's I Test: \code{\link{moranbi.test}}
  \item LISA Cluster and Significance Map: \code{\link{moran.cluster}}
  \item Create object "nb": \code{\link[spdep]{poly2nb}}
  \item Create object "listw"/"nb": \code{\link[spdep]{nb2listw}}
}
}
