% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranbi.test.R
\name{moranbi.test}
\alias{moranbi.test}
\title{Test for Bivariate Moran's I}
\usage{
moranbi.test(X, Y, listw, zero.policy = NULL, adjust.n = TRUE, N,
  graph = FALSE, print.results = TRUE, ...)
}
\arguments{
\item{X}{first variable}

\item{Y}{second variable}

\item{listw}{a neighbours list with spatial weights. From package spdep:
a listw object. Use poly2nb (class nb)
and nb2listw (class listw, nb) from package spdep. Can be any type of listw
object, for instance, rook contiguity (common edge) or queen contiguity (common
edge or common vertex)}

\item{zero.policy}{by default = NULL, if FALSE stop with error for any empty
neighbours sets, if TRUE permit the weights list to be formed with zero-length
weights vectors. Parameter inherited from the spdep package.}

\item{adjust.n}{by default = TRUE. if FALSE the number of observations is not
adjusted for noneighbour observations, if TRUE, the number of observations is
adjusted. Parameter inherited from the spdep package}

\item{N}{number of random elements for empirical density}

\item{graph}{by default = FALSE. Use TRUE to create test's graphic.}

\item{print.results}{by default = TRUE. Use FALSE to hide test results (table).
Results are: observed, expected and p-value.}

\item{...}{other parameters similar to original}
}
\value{
a list with three values: observed (Bivariate Moran's Index), expected
(-1/(N-1), and p-value.
}
\description{
Allow to test the null hyphothesis (ho) of zero spatial autocorrelation,
comparing p-value with a significance levels of alpha (0.1, or 0.05, or 0.01).
}
\details{
Compare the observed Bivariate Moran's I (moran.bi function) with the
expected value empirical density. The expected value is -1/(N-1), where N is
the number of rows/samples (number of polygons), and represents the null hyphothesis
(ho) of No Spatial Autocorrelation (random, observed Bivariate Moran's I around zero).
This expected value density is constructed with Monte Carlo simulations.
Values significant below of -1/(N-1) represents negative spatial autocorrelation
(generally negative values of observed bivariate Moran's I), and values
significant above of -1/(N-1) represents positive spatial autocorrelation
(generally positive values of observed bivariate Moran's I).
For hypothesis testing the sample values are compared with empirical density,
and p-value is calculated. Significant values of p-value (reject Ho) are smaller
than 10\% (if alpha = 10\%), smaller than 5\% (if alpha = 5\%) or smaller than 1\% (
if alpha = 1\%). For significant values of p-value (reject Ho), the conclusion
of the test could be: "given the value of p-value, there is less than alpha (1\%,
or 5\%, or 10\%) likelihood that the pattern (clustered or dispersed) could be the
result of random change".
}
\section{Links}{


\enumerate{
   \item Spatial Autocorrelation \href{http://blogs.oregonstate.edu/geo599spatialstatistics/2016/06/08/spatial-autocorrelation-morans/}{(Moran’s I) Test}
   \item Moran's I \href{https://stats.idre.ucla.edu/r/faq/how-can-i-calculate-morans-i-in-r/}{Test}
}
}

\examples{
library(spdep)
columbus <- st_read(system.file("extdata/columbus.shp", package="spdepbi")[1], quiet=TRUE)
col_nbq <- poly2nb(columbus)
a.lw <- nb2listw(col_nbq, style="W")
set.seed(123)
MBCrime <- moranbi.test(columbus$CRIME,columbus$INC,a.lw,N=999,graph=TRUE,zero.policy =TRUE)
moranbi.test(columbus$INC,columbus$HOVAL,a.lw,999,graph=TRUE,zero.policy=TRUE,N=1000)
}
\seealso{
\itemize{
  \item Bivariate Moran's I: \code{\link{moran.bi}}
  \item Plot Bivariate Moran's I: \code{\link{moranbi.plot}}
  \item Create object "nb": \code{\link[spdep]{poly2nb}}
  \item Create object "listw"/"nb": \code{\link[spdep]{nb2listw}}
}
}
